added a new weather + fixed int/long problem in abstract weather

+++ SmartGrid/src/main/java/com/rug/energygrid/weather/ExampleDataSet_KNI.java
@@ -5,10 +5,6 @@ import java.time.LocalDateTime;
 import java.time.ZoneId;
 import java.time.format.DateTimeFormatter;
 
-/**
- * Created by Ruben on 26-Apr-17.
- */
-
 //Static example, API would call specific dataSet in each method (This has to be implemented for
 public class ExampleDataSet_KNI extends Weather {
 
@@ -17,22 +13,21 @@ public class ExampleDataSet_KNI extends Weather {
         this.dataSet = readDataSet("weatherDataKNI.txt", ",", " ", "-"); //This is done because of Constant set //TODO: this can be speed up, now we read/parse everytime we create a document, but we could also read once and copy the data for each weather
     }
 
-    //These methods change depending on constant set or API
+    //These methods change depending on constants of the dataset(or depending on a API)
     protected Double getImpWindSpeed(Instant time) {
-        int iTime = convertToIntOfDataSet(time);
+        long iTime = convertToIntOfDataSet(time);
         return dataSet.get(iTime).get(WIND_SPEED_POS)*(0.1);
     }
 
     protected Double getImpSunIrradiation(Instant time) {
-        int iTime = convertToIntOfDataSet(time);
+        long iTime = convertToIntOfDataSet(time);
         return dataSet.get(iTime).get(SOLAR_IRRIDIANCE_POS);
     }
 
-    //when using data set has to be implemented
-    protected int convertToIntOfDataSet(Instant time) {
+    protected long convertToIntOfDataSet(Instant time) {
         LocalDateTime ldt = LocalDateTime.ofInstant(time, ZoneId.systemDefault());
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyyMMdd");
-        return Integer.valueOf(ldt.format(formatter));
+        return Long.valueOf(ldt.format(formatter));
     }
 
     @Override
